% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSDTreg.r
\name{coef.metaSDTreg}
\alias{coef.metaSDTreg}
\title{Coefficients method for metaSDTreg}
\usage{
\method{coef}{metaSDTreg}(object, ...)
}
\arguments{
\item{object}{An object of class metaSDTreg.}

\item{...}{For future methods.}
}
\value{
A named vector of parameter estimates.
}
\description{
Coefficients method for metaSDTreg
}
\examples{
## Declare simulated data as metaSDTdata
metadata <- metaSDTdata(simMetaData, type1='resp', type2='conf', signal='S')

## Fit model to subset of data
fit <- metaSDTreg(A ~ signal,
            data=metadata,
            subset = m <= 20)
coef(fit)

}
