% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSDTreg.r
\name{logLik.metaSDTreg}
\alias{logLik.metaSDTreg}
\title{Log-likelihood of metaSDTreg}
\usage{
\method{logLik}{metaSDTreg}(object, ...)
}
\arguments{
\item{object}{Object of class 'metaSDTreg'.}

\item{...}{For further methods.}
}
\value{
Numeric, the likelihood at the maximum found by the optimisation procedure.
}
\description{
Extract the log-likelihood from a metaSDTreg object.
}
\examples{
## Declare simulated data as metaSDTdata
metadata <- metaSDTdata(simMetaData, type1='resp', type2='conf', signal='S')

## Fit model to subset of data
fit <- metaSDTreg(A ~ signal,
            data=metadata,
            subset = m <= 20)
logLik(fit)

}
