% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSDTreg.r
\name{metaSDT.loglik}
\alias{metaSDT.loglik}
\title{Meta-SDT log-likelihood}
\usage{
metaSDT.loglik(parm, df, A, X, X.n = NULL, X.s = NULL, L0, K0)
}
\arguments{
\item{parm}{Parameters of the model, c(intercepts^T,d.n,d,d.s,beta^T)^T, where length(intercepts)=K0-1 and lenght(beta)=ncol(X).}

\item{df}{Data frame}

\item{A}{Matrix of responses}

\item{X}{Matrix of covariates}

\item{X.n}{Presently not in use}

\item{X.s}{Presently not in use}

\item{L0}{Levels of the type 2 task}

\item{K0}{Levels of the ordinal variable A}
}
\description{
Meta SDT log-likelihood function, internal function. Used by \code{metaSDTreg}.
}
\keyword{internal}
