% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_roc.r
\name{predict_roc}
\alias{predict_roc}
\title{Generic predict_roc method}
\usage{
predict_roc(object, ...)
}
\arguments{
\item{object}{Signal detection theory model.}

\item{...}{Further arguments passed to methods.}
}
\value{
An object of class 'predict_roc'.
}
\description{
Predict ROC curve from signal detection theory model.
}
\seealso{
\code{\link{predict_roc.metaSDTreg}}, \code{\link{predict_roc.metaSDTdata}}.
}
