% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSDTdata.r
\name{print.metaSDTdata}
\alias{print.metaSDTdata}
\title{Print method for metaSDT data}
\usage{
\method{print}{metaSDTdata}(x, ...)
}
\arguments{
\item{x}{The metaSDTdata object.}

\item{...}{Additional arguments passed to print.data.frame}
}
\value{
Invisible.
}
\description{
Print method for metaSDT data
}
\examples{
metadata <- metaSDTdata(simMetaData, type1='resp', type2='conf', signal='S')
print(metadata)

}
