% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSDTreg.r
\name{print.summary.metaSDTreg}
\alias{print.summary.metaSDTreg}
\title{Print summary method for metaSDTreg}
\usage{
\method{print}{summary.metaSDTreg}(x, ...)
}
\arguments{
\item{x}{An object of class summary.metaSDTreg.}

\item{...}{For future methods.}
}
\value{
Invisible.
}
\description{
Print summary method for metaSDTreg
}
\examples{
## Declare simulated data as metaSDTdata
metadata <- metaSDTdata(simMetaData, type1='resp', type2='conf', signal='S')

## Fit model to subset of data
fit <- metaSDTreg(A ~ signal,
            data=metadata,
            subset = m <= 20)
summary(fit)

}
