% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_roc.r
\name{ROCcoords}
\alias{ROCcoords}
\title{Extract coordinates for predicted ROC curve}
\usage{
ROCcoords(x, thr)
}
\arguments{
\item{x}{A 'predict_roc' object to plot.}

\item{thr}{The sequence of thresholds parametrising the ROC curve, if this is a function.}
}
\description{
Extract coordinates for predicted ROC curve
}
\keyword{internal}
