% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSDTreg.r
\docType{data}
\name{simMetaData}
\alias{simMetaData}
\title{Simulated metacognition experiment}
\format{
A data frame with 25000 obs. of  5 variables:
\describe{
  \item{id}{Subject id.}
  \item{m}{Observation id, within subject.}
  \item{S}{Signal indicator (0 or 1).}
  \item{resp}{Type 1 response (0 or 1).}
  \item{conf}{Type 2 response, ordered (1, 2 or 3).} 
}
}
\usage{
simMetaData
}
\description{
Data is simulated from the meta-SDT model, formula (3)-(4) in Kristensen, Sandberg and Bibby (2020).
The true values of the simulation parameters are
c(tau.n.2=-0.5, tau.n.1=-0.2, theta=0,
tau.s.1=0.8, tau.s.2=1, d.n=0.5, d=1, d.s=0.75).
}
\references{
Kristensen, S. B., Sandberg, K., & Bibby, B. M. (2020). Regression
methods for metacognitive sensitivity. Journal of Mathematical
Psychology, 94. <doi:10.1016/j.jmp.2019.102297>.
}
\keyword{datasets}
