% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSDTdata.r
\name{summary.metaSDTdata}
\alias{summary.metaSDTdata}
\title{Summarise a metaSDTdata object as a cognitive experiment.}
\usage{
\method{summary}{metaSDTdata}(object, ...)
}
\arguments{
\item{object}{The metaSDTdata object.}

\item{...}{Additional arguments. Presently not used.}
}
\value{
A list with class 'summary.metaSDTdata' containing summaries.
}
\description{
Summarise a metaSDTdata object as a cognitive experiment.
}
\examples{
metadata <- metaSDTdata(simMetaData, type1='resp', type2='conf', signal='S')
summary(metadata)

}
