% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSDTreg.r
\name{vcov.metaSDTreg}
\alias{vcov.metaSDTreg}
\title{Variance-covariance method for metaSDTreg}
\usage{
\method{vcov}{metaSDTreg}(object, ...)
}
\arguments{
\item{object}{An object of class metaSDTreg.}

\item{...}{For future methods.}
}
\value{
A matrix of variances and covariances.
}
\description{
Variance-covariance method for metaSDTreg
}
\examples{
## Declare simulated data as metaSDTdata
metadata <- metaSDTdata(simMetaData, type1='resp', type2='conf', signal='S')

## Fit model to subset of data
fit <- metaSDTreg(A ~ signal,
            data=metadata,
            subset = m <= 20)

## Standard errors
sqrt(diag(vcov(fit)))


}
